pyproject.toml
[project]
name = "promptlightning"
version = "1.0.2"
description = "A Python library for managing and rendering prompt templates with type-safe inputs, versioning, and optional logging"
readme = "README.md"
authors = [
    { name = "bogdanpi", email = "bogdan@iduntechnologies.com" }
]
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.2",
    "typer>=0.19.2",
    "watchdog>=6.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "lmdb>=1.4.0",
    "msgpack>=1.0.5",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
promptlightning = "promptlightning.cli:app"

[project.urls]
Homepage = "https://github.com/bogdan-pistol/promptlightning"
Repository = "https://github.com/bogdan-pistol/promptlightning.git"
Issues = "https://github.com/bogdan-pistol/promptlightning/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["promptlightning"]

[tool.hatch.build.targets.wheel.force-include]
# Include playground directory (built by CI, or empty with .gitkeep)
"playground" = "playground"

[tool.hatch.build.targets.sdist]
# Source dist includes promptlightning source and web/ source
exclude = ["web/node_modules"]

[tool.hatch.build.targets.sdist.force-include]
# Force include playground directory structure
"playground" = "playground"


[tool.uv]
dev-dependencies = [
    "build>=1.3.0",
    "twine>=6.2.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "tomli-w>=1.0.0",
]
